==> R/2-muir-analysis.Rmd <==
---
title: "Re-analysis of Muir et al. data"
author: "Eric R. Scott"
output: 
  html_notebook:
    code_folding: hide
    number_sections: yes
    theme: paper
---

```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(ropls)
library(glue)
library(colorspace)
library(ggrepel)
library(lmtest)
library(broom)
library(patchwork)

source(here("R", "ropls_helpers.R"))
source(here("R", "ropls_plots.R"))
```

# Read in Data

I'm going to remove the outlier species, *S. juglandifolium* for the sake of this analysis

```{r}
tomato <- read_rds(here("data", "tomatoes.rds"))

#remove S.juglandifolium
tomato.filtered <-
  tomato %>% 
  filter(Spe != "jug") %>% 
  mutate(temp = temp * 0.1) %>% 
  filter(complete.cases(.))
```


# PCA regression

```{r}
X <-tomato.filtered %>% dplyr::select(lma.leaflet:Rdark)
tom.pca <- opls(X, plotL = FALSE)
```

Precipitation, is not significantly related to PC1 (i.e. the LES)

```{r}
tom.pca.data <- get_plotdata(tom.pca)
pcr.df2 <-
  tom.pca.data$scores %>% 
  add_column(precip = tomato.filtered$precip)
```

```{r}
m0 <- lm(precip ~ 1, data = pcr.df2)
m1 <- lm(precip ~ p1, data = pcr.df2)

lrtest(m0, m1)
summary(m1)
```


## Biplot

Combines scores and loadings. I'm specifically making a *column-principal biplot* where the scores are in standard coordinates (sums of squares = 1 on each axis) and the loadings are in principal coordinates (sums of squares = variance explained by that axis).  This is according to

Greenacre, M. & Primicerio, R. (2013) Ch12. Principal Component Analysis. In *Multivariate Analysis of Ecological Data*. pp. 151–161.


```{r}
scaled_loadings <- get_loadings(tom.pca) %>% 
  mutate(p1 = ss_scale(p1, tom.pca.data$axis_stats$R2X[1]),
         p2 = ss_scale(p2, tom.pca.data$axis_stats$R2X[2])) %>% 
  mutate(Variable = case_when(
    Variable == "lma.leaflet" ~ "LMA",
    Variable == "thick.vile" ~ "Thickness",
    Variable == "ldmc" ~ "LDMC",
    Variable == "wue" ~ "WUE",
    TRUE ~ Variable
  ))

scaled_scores <- get_scores(tom.pca) %>% 
  mutate_if(is.numeric, ss_scale) %>% 
  add_column(precip = tomato.filtered$precip)


xlab <- TeX(glue("Comp. 1 ($R^2_X = {tom.pca.data$axis_stats[1,1] * 100}%$, $R^2_Y = {round(glance(m1)$r.squared * 100, 1)}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {tom.pca.data$axis_stats[2,1] * 100}%$)"))

pca.biplot <- 
  ggplot(scaled_scores) +
  geom_point(aes(x = p1, y = p2, fill = precip),
             shape = 21, size = 2) +
  scale_fill_viridis_b(guide = guide_bins(reverse = TRUE, show.limits = TRUE, title.vjust = 2)) +
  geom_segment(data = scaled_loadings, 
               aes(x = 0, y = 0, xend = p1, yend = p2),
               alpha = 0.6,
               arrow = arrow(length = unit(0.15, "cm"))) +
  geom_label_repel(data = scaled_loadings, 
                   aes(x = p1, y = p2, label = Variable),
                   segment.alpha = 0.6, direction = "y", size = 2.7, 
                   min.segment.length = 5, force = 10) +
  labs(x = xlab,
       y = ylab,
       fill = "Annual Precipitation (mm)",
       subtitle = "a) Principal Components Analysis (PCA)") +
  theme_bw()
pca.biplot
```


# PLSR

Precipitation is significantly related to leaf traits.  The predictive axes explain very little of the variation in the leaf traits (`R2X(cum)`), but a significant amount of variation in precipitation (`R2Y(cum)` and `pR2Y` and `pQ2`).  `Q2(cum)` is not much lower than `R2Y(cum)` indicating an absence of overfitting or a predictive power similar to the explanatory power.

```{r}
tom.pls <- 
  opls(X, tomato.filtered$precip,
       fig.pdfC = "none",
       permI = 1000)
#produces a single component model.  I will "force" two predictive axes for the sake of plotting.
tom.pls2 <- 
  opls(X, tomato.filtered$precip,
       fig.pdfC = "none", 
       predI = 2, permI = 1000)

get_modelinfo(tom.pls)
```


## PLSR bi-plot

```{r}
tom.pls.data <- get_modelinfo(tom.pls2)
scaled_loadings <- get_loadings(tom.pls2) %>% 
  mutate(p1 = ss_scale(p1, tom.pls.data$axis_stats$R2X[1]),
         p2 = ss_scale(p2, tom.pls.data$axis_stats$R2X[2])) %>% 
    mutate(Variable = case_when(
    Variable == "lma.leaflet" ~ "LMA",
    Variable == "thick.vile" ~ "Thickness",
    Variable == "ldmc" ~ "LDMC",
    Variable == "wue" ~ "WUE",
    TRUE ~ Variable
  ))

scaled_scores <- get_scores(tom.pls2) %>% 
  mutate_if(is.numeric, ss_scale) %>% 
  add_column(precip = tomato.filtered$precip)

xlab <- TeX(glue("Comp. 1 ($R^2_X = {tom.pls.data$axis_stats[1,1] * 100}%$, $R^2_Y = {tom.pls.data$axis_stats[1,3] * 100}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {tom.pls.data$axis_stats[2,1] * 100}%$, $R^2_Y = {tom.pls.data$axis_stats[2,3] * 100}%$)"))
pls.biplot <- 
  ggplot(scaled_scores) +
  geom_point(aes(x = p1, y = p2, fill = precip),
             shape = 21, size = 2) +
  scale_fill_viridis_b(guide = guide_bins(reverse = TRUE, show.limits = TRUE, title.vjust = 2)) +
  geom_segment(data = scaled_loadings,
               aes(x = 0, y = 0, xend = p1, yend = p2),
               alpha = 0.6,
               arrow = arrow(length = unit(0.15, "cm"))) +
  geom_label_repel(data = scaled_loadings,
                   aes(x = p1, y = p2, label = Variable),
                   segment.alpha = 0.6, direction = "y", size = 2.7, 
                   min.segment.length = 5, force = 10) +
  labs(x = xlab,
       y = ylab,
       fill = "Annual Precipitation (mm)",
       subtitle = "b) Partial Least Squares (PLS) Regression") +
  theme_bw()
pls.biplot
```
## Compose and output plot

Oecologia wants panel labels *inside* of plot borders and a "Fig. 2" label outside of the panel borders.

```{r}
patchwork <- 
  pca.biplot + pls.biplot  +
  plot_layout(guides = "collect") &
  theme_classic() &
  theme(panel.grid = element_blank(),
        panel.border = element_blank()) &
  labs(fill = "Precip.\n(mm/yr)")
```


```{r}
patchwork2 <-
  patchwork + 
  #for Oecologia final submission:
  plot_annotation(title = "Fig. 3", tag_levels = "a", tag_suffix = ")") &
  labs(subtitle = NULL) & #temporary.  want to keep these, but have them inside of plot boundaries.
  theme(
    plot.tag.position = c(0,1),
    plot.tag = element_text(vjust = 1.5, hjust = -4)
  ) 


patchwork2
ggsave(here("out", "muir-biplots.png"), width = 9, height = 4)
ggsave(here("out", "muir-biplots.pdf"), width = 9, height = 4)
```


# Compare Loadings

Table of loadings from the first PCA and PLS axes.

```{r}
pls.loadings <- get_loadings(tom.pls) %>% select(Variable, PLS = p1)
pca.loadings <- get_loadings(tom.pca) %>% select(Variable, PCA = p1)

loading.table <-
  full_join(pca.loadings, pls.loadings) %>%
  mutate_if(is.numeric, ~round(., 3))

loading.table
write_csv(loading.table, here("out", "muir loading table.csv"))
```
# Correlations

Make a table of axis correlations instead of loadings.  I think these are mathematically related.

According to Cadima & Joliffe 1995:

$$
\rho_{ij} = a_{ij}\bigg(\frac{\lambda_i}{s_{ij}}\bigg)^{1/2}
$$

Where $a_{ij}$ is the loading for the $i$th variable in the $j$the PC, $\lambda$ is the eigenvalue of the PC, and $s_{ij}$ is the variance of the variable, which should be 1 because I scaled everything to unit variance.  So, it simplifies to:

$$
\rho_{ij} = a_{ij}\times \sqrt{\lambda_i}
$$
See also: https://stats.stackexchange.com/questions/143905/loadings-vs-eigenvectors-in-pca-when-to-use-one-or-another

But, I'm just going to calculate this as the correlation between the scores and the data.


```{r}
pc1 <- get_scores(tom.pca)$p1
p1 <- get_scores(tom.pls)$p1

cor.table <- 
  tibble(Variable = loading.table$Variable,
       PCA = map_dbl(X, ~cor(.x, pc1)),
       PLS = map_dbl(X, ~cor(.x, p1)))
cor.table
write_csv(cor.table, here("out", "muir correlation table.csv"))
```

# Plots for ESA talk

```{r}
pca.biplot +
  labs(x = "PC1 (31.9%)", y = "PC2 (24.2%)", fill = "Precip.\n(mm/yr)") +
  theme(text = element_text(size = 14))
ggsave(here("out", "muir_pca.png"), height = 4, width = 5)
```

```{r}
pls.biplot +
  labs(x = "Predictive Comp. 1 (19.5%)", y = "Predictive Comp. 2 (25.4%)", fill = "Precip.\n(mm/yr)") +
  theme(text = element_text(size = 14))
ggsave(here("out", "muir_pls.png"), height = 4, width = 5)
```

```{r}

bind_cols(tomato.filtered, get_scores(tom.pca)) %>% 
  arrange(ldmc)
```
