==> R/3-generating-datasets.Rmd <==
---
title: "Simulating Datasets"
author: "Eric R. Scott"
output: html_notebook
---
```{r message=FALSE, warning=FALSE}
library(iheatmapr)
library(holodeck)
library(tidyverse)
library(here)
library(furrr)
myseed <- 258
#set up parallelization
plan(multiprocess)
```

**IMPORTANT:** These datasets were generated under R version 3.6.1.  The random number generator in R was updated in version 3.6.0.  For this to be fully reproducible, run with version > 3.6.0.

**NOTE:** This script used the `furrr` package which adds parallelization support to `map()` and other functions from the `purrr` package.  If your machine doesn't support parallelization, or you run into problems, you can replace `future_map()` with `map()`.

**For all datasets:**

 - n_vars (number of variables) = 25
 - n_obs (number of observations) = 20
 - var (variation parameter, passed to mvrnorm) = 1
 - Number of groups = 2
 - Number of datasets generated = 500
 
**Scenarios:**
1. "No predictors"
    - no discriminating variables
    - 25 variables with cov = 0

2. "Apparent predictors" (i.e. correlated variables are also predictive)
    - 5 variables with cov = 0
    - two groups of 10 variables that discriminate between groups with difference in means of 2 and cov = 0.5 

3. "Hidden Predictors" (i.e. predictive variables are not correlated and not the main axis of variation)
    - 5 variables that discriminate between groups with a difference in means of 2, cov = 0
    - two groups of 10 variables with cov = 0.5


# Simulate Data

## 1. "No Predictors"

```{r}
set.seed(myseed)
none <-
  future_map(1:500,
      ~ sim_cat(n_obs = 20, n_groups = 2) %>% 
        sim_covar(n_vars = 25, var = 1, cov = 0, name = "N")
  ) %>%
  set_names(paste0("none_df_", 1:500))
```

Inspect one:

```{r}
set.seed(myseed)
none[[sample(1:500, 1)]] %>%
  dplyr::select(-group) %>% 
  cor() %>% 
  iheatmap(row_labels = TRUE, col_labels = TRUE, name = "cor")
```


## 2. Apparent Predictors

```{r}
set.seed(myseed)

apparent <- 
  future_map(1:500,
             ~sim_cat(n_obs = 20, n_groups = 2) %>% 
               sim_covar(n_vars = 5, var = 1, cov = 0, name = "N") %>% 
               group_by(group) %>% 
               sim_discr(n_vars = 10, var = 1, cov = 0.5, group_means = c(-1, 1), name = "D1") %>% 
               sim_discr(n_vars = 10, var = 1, cov = 0.5, group_means = c(1, -1), name = "D2") %>% 
               ungroup()
  ) %>% 
  set_names(paste0("apparent_df_", 1:500))
```

Inspect one:

```{r}
set.seed(myseed)
apparent[[sample(1:500, 1)]] %>%
  dplyr::select(-group) %>% 
  cor() %>% 
  iheatmap(row_labels = TRUE, col_labels = TRUE, name = "cor")
```

## 3. "Hidden Predictors:

```{r}
set.seed(myseed)
hidden <-
  future_map(1:500,
      ~ sim_cat(n_obs = 20, n_groups = 2) %>% 
        sim_covar(n_vars = 10, var = 1, cov = 0.5, name = "C1") %>% 
        sim_covar(n_vars = 10, var = 1, cov = 0.5, name = "C2") %>% 
        group_by(group) %>% 
        sim_discr(n_vars = 5, var = 1, cov = 0, group_means = c(-1, 1), name = "D") %>% 
        ungroup()
  ) %>%
  set_names(paste0("hidden", 1:500))
```

Inspect one:

```{r}
hidden[[sample(1:500, 1)]] %>%
  dplyr::select(-group) %>% 
  cor() %>% 
  iheatmap(row_labels = TRUE, col_labels = TRUE, name = "cor")
```


```{r}
write_rds(none, here("data", "none.rds"))
write_rds(hidden, here("data", "hidden.rds"))
write_rds(apparent, here("data", "apparent.rds"))
```
