==> R/4-analyze-simulated-data.Rmd <==
---
title: "Simulated data analysis"
author: "Eric R. Scott"
output: 
  html_notebook: 
    toc: yes
    toc_float: yes
---

# Overview
For each simulated dataset in each of the three scenarios:

- Do PCA logistic regression
- Do PLS-DA
- Store results
  
I'll do this both forcing 2 axes (predI = 2) as well as letting `opls()` decide how many axes to retain.

# Packages and Functions

```{r message=FALSE, warning=FALSE}
library(here)
library(ropls)
library(tidyverse)
library(furrr) #for parallelization
source(here("R","pcr.R"))
library(beepr) #for alerts when long-running code is finished.
set.seed(690)
```

Set up parallelization

```{r include=FALSE}
plan(multiprocess)
```


Create "safe" "partial" version of `opls()` which returns `NULL` if it errors and never plots or prints results.

```{r}
safe_opls <- safely(partial(opls, info.txtC = "none", fig.pdfC = "none"), otherwise = NULL)
safe_pcr <- safely(pca_lr, otherwise = NULL)
```

# Load Data

```{r}
none <- read_rds(here("data", "none.rds"))
hidden <- read_rds(here("data", "hidden.rds"))
apparent <- read_rds(here("data", "apparent.rds"))
```


# "No Predictors"

## "No Predictors" PCAs

```{r}
none.pcr.try <- 
  none %>% 
  future_map(~safe_pcr(.x, -group, group, CV = 7, predI = 2))
# Get just the successful models
none.pcr <-
  none.pcr.try %>%
  future_map("result") %>% #get only results
  discard(is.null) #get rid of any with NULL results (and therefore an error)
beep(4)
```

Any errors?

```{r}
length(none.pcr.try) - length(none.pcr)
none.pcr.try %>% 
  map("error") %>%
  unique()
```

2 errors in PCA part, but also several convergence warnings in GLM part.

### Allowing opls to choose ncomp

```{r}
none.pcr.auto.try <- 
  none %>% 
  future_map(~safe_pcr(.x, -group, group, CV = 7))
# Get just the successful models
none.pcr.auto <-
  none.pcr.auto.try %>%
  future_map("result") %>% #get only results
  discard(is.null) #get rid of any with NULL results (and therefore an error)
beep(4)
```


Save em!

```{r}
write_rds(none.pcr, here("data", "models", "none-pcr.rds"))
write_rds(none.pcr.auto, here("data", "models", "none-pcr-auto.rds"))
```


## "No Predictors" PLSs

```{r}
none.pls.try <-
  none %>% 
  future_map(~ safe_opls(select(., -group), .$group, predI = 2, permI = 500)) #map the opls

# get just succesful models
none.pls <-
  none.pls.try %>% 
  future_map("result") %>% 
  discard(is.null) 
beep(4)
```

Any errors?

```{r}
length(none.pls.try) - length(none.pls)
```

None (because forced 2 components)

### Allowing opls to choose number of components

```{r}
none.pls.auto.try <-
  none %>% 
  future_map(~ safe_opls(select(., -group), .$group, permI = 500)) #map the opls

# get just succesful models
none.pls.auto <-
  none.pls.auto.try %>% 
  future_map("result") %>% 
  discard(is.null) 
beep(4)
```

```{r}
write_rds(none.pls, here("data", "models", "none-pls.rds"))
write_rds(none.pls.auto, here("data", "models", "none-pls-auto.rds"))
```

# "Apparent Predictors:

## "Apparent Predictors" PCAs

```{r message=FALSE, warning=FALSE}
# map the pca function to all datasets
apparent.pcr.try <-
  future_map(apparent,
      ~safe_pcr(.x, -group, group, CV = 7, predI = 2))

apparent.pcr <-
  apparent.pcr.try %>%
  future_map("result") %>% 
  discard(is.null)
beep(4)
```

This gives a bunch of convergence warnings, but it's just because fitted values are perfectly 0 or 1, which is safe to ignore (I believe).

How many errors? 

```{r}
length(apparent.pcr.try) - length(apparent.pcr)
```
no errors.

### Allowing opls to decide how many components

```{r message=FALSE, warning=FALSE}
# map the pca function to all datasets
apparent.pcr.auto.try <-
  future_map(apparent,
      ~safe_pcr(.x, -group, group, CV = 7))

apparent.pcr.auto <-
  apparent.pcr.auto.try %>%
  future_map("result") %>% 
  discard(is.null)
beep(4)
```

Save em!

```{r}
write_rds(apparent.pcr, here("data", "models", "apparent-pcr.rds"))
write_rds(apparent.pcr.auto, here("data", "models", "apparent-pcr-auto.rds"))
```


## "Apparent Predictors: PLS-DAs

```{r}
apparent.pls.try <-
  future_map(apparent,
      ~ safe_opls(select(., -group), .$group, predI = 2, permI = 500))

apparent.pls <-
  apparent.pls.try %>% 
  future_map("result") %>% 
  discard(is.null)
beep(4)
```

How many failed models?

```{r}
length(apparent.pls.try) - length(apparent.pls)
```

zero failed models
### Allowing opls to decide how many components

```{r}
apparent.pls.auto.try <-
  future_map(apparent,
      ~ safe_opls(select(., -group), .$group, permI = 500))

apparent.pls.auto <-
  apparent.pls.auto.try %>% 
  future_map("result") %>% 
  discard(is.null)
beep(4)
```
Save em!

```{r}
write_rds(apparent.pls, here("data", "models", "apparent-pls.rds"))
write_rds(apparent.pls.auto, here("data", "models", "apparent-pls-auto.rds"))
```

# "Hidden Predictors"

## "Hidden Predictors" PCAs

```{r message=FALSE, warning=FALSE}
hidden.pcr.try <-
  future_map(hidden,
      ~safe_pcr(.x, -group, group, predI = 2, CV = 7))

hidden.pcr <-
  hidden.pcr.try %>% 
  future_map("result") %>% 
  discard(is.null)
beep(4)
```

Any errors?

```{r}
length(hidden.pcr.try) - length(hidden.pcr)
#What was the error?
hidden.pcr.try %>% 
  future_map("error") %>% 
  unique()
```

2 failed models due to the PCA part failing (exactly singular error)

### Allowing opls to decide how many components

```{r message=FALSE, warning=FALSE}
hidden.pcr.auto.try <-
  future_map(hidden,
      ~safe_pcr(.x, -group, group, CV = 7))

hidden.pcr.auto <-
  hidden.pcr.auto.try %>% 
  future_map("result") %>% 
  discard(is.null)
beep(4)
```


```{r}
write_rds(hidden.pcr, here("data", "models", "hidden-pcr.rds"))
write_rds(hidden.pcr.auto, here("data", "models", "hidden-pcr-auto.rds"))
```

## Needle PLS-DAs

```{r}
hidden.pls.try <-
  future_map(hidden,
      ~ safe_opls(select(., -group), .$group, predI = 2, permI = 500))

hidden.pls <-
  hidden.pls.try %>% 
  future_map("result") %>% 
  discard(is.null)
beep(4)
```

Any errors?

```{r}
length(hidden.pls.try) - length(hidden.pls)
```

zero errors

```{r}
hidden.pls.auto.try <-
  future_map(hidden,
      ~ safe_opls(select(., -group), .$group, permI = 500))

hidden.pls.auto <-
  hidden.pls.auto.try %>% 
  future_map("result") %>% 
  discard(is.null)
beep(4)
```

Save 'em!

```{r}
write_rds(hidden.pls, here("data", "models", "hidden-pls.rds"))
write_rds(hidden.pls.auto, here("data", "models", "hidden-pls-auto.rds"))
```
