==> R/6-fig2-pca-vs-pls.Rmd <==
---
title: "PCA vs. PLS-DA"
author: "Eric R. Scott"
output: html_notebook
---

This generates figure 2 showing the correlation structure and results from PCA-LR and PLS-DA for example datasets from the three scenarios.

**IMPORTANT:** The random number generator in R was updated in version 3.6.0.  For this to be fully reproducible, run with version > 3.6.0

```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(ropls) #for opls(), which does pca and pls-da
library(cowplot) #for plot_grid()
theme_set(theme_gray())
library(here)
library(colorspace)
source(here("R", "ropls_plots.R"))
source(here("R", "ropls_helpers.R"))
source(here("R", "pcr.R"))
library(glue)
library(latex2exp)
```


# Read in datasets
I'll read in the datasets from "3-generating-datasets.Rmd" and pick individual datasets to make figures

```{r}
null.list <- read_rds(here::here("data", "null.rds"))
control.list <- read_rds(here::here("data", "control.rds"))
needle.list <- read_rds(here::here("data", "needle.rds"))
herring.list <- read_rds(here::here("data", "red_herring.rds"))
```

Sample one dataset as an example

```{r}
# set.seed(719)
set.seed(720)
null <- sample(null.list, 1)[[1]]
needle <- sample(needle.list, 1)[[1]]
control <- sample(control.list, 1)[[1]]
herring <- sample(herring.list, 1)[[1]]
```

# Null

## Correlation heatmap for Null

```{r}
null.cor <-
  null %>% 
  select(-group) %>% 
  cor() %>%
  as_tibble(rownames = "row") %>% 
  gather(-row, key = col, value = cor)

null.cor.p <-
  ggplot(null.cor, aes(x = col, y = row, fill = cor)) +
  geom_tile() +
  scale_fill_continuous_diverging(palette = "Blue-Red 3", limits = c(-1,1)) +
  labs(x = NULL, y = NULL, subtitle = '"No predictors"', fill = "r") +
  theme_minimal() +
  theme(legend.position = "right")

# Replace axis labels with curly braces to improve readability:
b <- bracketsGrob(0.995, 0, 0.005, 0, h = 0.05)

bracket_labels <- tibble(x = 13, y = -1.75, label = "N")

null.cor.p <-
  null.cor.p +
  coord_cartesian(ylim = c(1,25), clip = "off") + #allows brackets to show up off plot
  theme(
    axis.text = element_blank(), #turns off axis labels
    axis.ticks = element_blank(), #turns off axis ticks
    plot.margin = unit(c(1,1,1.5,1), "lines") #padds margin to make room
  ) +
  annotation_custom(b) +
  geom_text(data = bracket_labels, inherit.aes = FALSE, aes(x=x, y=y, label = label))

null.cor.p
```


## PCA and PLS-DA for Null

```{r}
null.pcalr <- pca_lr(null, -group, group, predI = 2, CV = 7)
null.pca <- null.pcalr$pca
null.pls <- opls(select(null, -group), null$group, fig.pdfC = "none", permI = 500, predI = 2)
```

## Create plots for Null

```{r message=FALSE, warning=FALSE}
null.pca.p <-
  plot_pca(null.pca, null$group, annotate = "subtitle") +
  labs(title = NULL,
       subtitle =TeX(glue("$R^2_Y = {signif(null.pcalr$mod.stats$R2Y, 2)}$; $p = {signif(null.pcalr$mod.stats$p.value, 2)}$"))) +
  scale_color_discrete_qualitative(palette = "Dark 3")

null.pls.p <-
  plot_plsda(null.pls, annotate = "subtitle") +
  scale_color_discrete_qualitative(palette = "Dark 3")
null.pca.p
null.pls.p
```


# Needle in a haystack

## Correlation heatmap for needle-in-a-haystack

```{r}
needle.cor <-
  needle %>% 
  select(-group) %>% 
  cor() %>%
  as_tibble(rownames = "row") %>%  
  gather(-row, key = col, value = cor)

needle.cor.p <-
  ggplot(needle.cor, aes(x = col, y = row, fill = cor)) +
  geom_tile() +
  scale_fill_continuous_diverging(palette = "Blue-Red 3", limits = c(-1,1)) +
  labs(x = NULL, y = NULL, subtitle = '"Hidden predictors"', fill = "r") +
  theme_minimal() +
  theme(legend.position = "right")

b1 <- bracketsGrob(10/25, 0, 0.005, 0, h = 0.05)
b2 <- bracketsGrob(20/25, 0, 10/25, 0, h = 0.05)
b3 <- bracketsGrob(0.995, 0, 20/25, 0, h = 0.05)

bracket_labels <- tibble(x = c(5.5, 15.5, 23), #where to put labels on x-axis
                         y = rep(-1.75, 3), #how far below bracket to place labels
                         label = c("C1", "C2", "D")) #labels for groups of variables

needle.cor.p <-
  needle.cor.p +
  coord_cartesian(ylim = c(1,25), clip = "off") + #allows brackets to show up off plot
  theme(
    axis.text = element_blank(), #turns off axis labels
    axis.ticks = element_blank(), #turns off axis ticks
    plot.margin = unit(c(1,1,1.5,1), "lines") #pads margin to make room
  ) +
  annotation_custom(b1) +
  annotation_custom(b2) +
  annotation_custom(b3) +
  geom_text(data = bracket_labels, inherit.aes = FALSE, aes(x=x, y=y, label = label))
needle.cor.p
```

## PCA and PLS-DA for needle-in-a-haystack

The PCA score plot still shows no real separation because the discriminating variables do not contribute much to the total covariation in the data.  The PLS-DA plot shows very strong separation, despite the differences between the groups being only due to 5 out of 25 variables.  The PLS-DA model performs well, with high R^2^ and Q^2^ values

```{r include=FALSE}
needle.pcalr <- pca_lr(needle, -group, group, predI = 2, CV = 7)
needle.pca <- needle.pcalr$pca
needle.pls <- opls(select(needle, -group), needle$group, fig.pdfC = "none", predI = 2, permI = 500)
```

## Plots for needle-in-a-haystack

```{r message=FALSE, warning=FALSE}
needle.pca.p <-
  plot_pca(needle.pca, needle$group, annotate = "subtitle") +
  labs(title = NULL,
       subtitle =TeX(glue("$R^2_Y = {signif(needle.pcalr$mod.stats$R2Y, 2)}$; $p = {signif(needle.pcalr$mod.stats$p.value, 2)}$"))) +
  scale_color_discrete_qualitative(palette = "Dark 3")

needle.pls.p <-
  plot_plsda(needle.pls, annotate = "subtitle") +
  scale_color_discrete_qualitative(palette = "Dark 3")
needle.pca.p
needle.pls.p
```


# Control

## Correlation heatmap for Control

```{r}
control.cor <-
  control %>% 
  select(-group) %>% 
  cor() %>%
  as_tibble(rownames = "row") %>% 
  gather(-row, key = col, value = cor)

control.cor.p <-
  ggplot(control.cor, aes(x = col, y = row, fill = cor)) +
  geom_tile() +
  scale_fill_continuous_diverging(palette = "Blue-Red 3", limits = c(-1,1)) +
  labs(x = NULL, y = NULL, subtitle = '"Apparent predictors"', fill = "r") +
  theme_minimal()

b1 <- bracketsGrob(10/25, 0, 0.005, 0, h = 0.05)
b2 <- bracketsGrob(20/25, 0, 10/25, 0, h = 0.05)
b3 <- bracketsGrob(0.995, 0, 20/25, 0, h = 0.05)

bracket_labels <- tibble(x = c(5.5, 15.5, 23), #where to put labels on x-axis
                         y = rep(-1.75, 3), #how far below bracket to place labels
                         label = c("D1", "D2", "N")) #labels for groups of variables
control.cor.p <-
  control.cor.p +
  coord_cartesian(ylim = c(1,25), clip = "off") + #allows brackets to show up off plot
  theme(
    axis.text = element_blank(), #turns off axis labels
    axis.ticks = element_blank(), #turns off axis ticks
    plot.margin = unit(c(1,1,1.5,1), "lines") #padds margin to make room
  ) +
  annotation_custom(b1) +
  annotation_custom(b2) +
  annotation_custom(b3) +
  geom_text(data = bracket_labels, inherit.aes = FALSE, aes(x=x, y=y, label = label))
control.cor.p
```

## PCA and PLS-DA for Control

```{r include=FALSE}
control.pcalr <- pca_lr(control, -group, group,  CV = 7, predI = 2) 

control.pca <- control.pcalr$pca
control.pls <- opls(select(control, -group), control$group, fig.pdfC = "none", predI = 2, permI = 500)
```

## Plots for Control

```{r message=FALSE, warning=FALSE}
control.pca.p <-
  plot_pca(control.pca, control$group, annotate = "subtitle") +
  labs(title = NULL,
       subtitle =TeX(glue("$R^2_Y = 0.99$; $p < 0.001$"))) +
  scale_color_discrete_qualitative(palette = "Dark 3")

control.pls.p <-
  plot_plsda(control.pls, annotate = "subtitle") +
  scale_color_discrete_qualitative(palette = "Dark 3")
control.pca.p
control.pls.p
```
# Red Herring

## Correlation heatmap for Red Herring

```{r}
herring.cor <-
  herring %>% 
  select(-group) %>% 
  cor() %>%
  as_tibble(rownames = "row") %>% 
  gather(-row, key = col, value = cor)

herring.cor.p <-
  ggplot(herring.cor, aes(x = col, y = row, fill = cor)) +
  geom_tile() +
  scale_fill_continuous_diverging(palette = "Blue-Red 3", limits = c(-1,1)) +
  labs(x = NULL, y = NULL, subtitle = '"Red Herring"', fill = "r") +
  theme_minimal()

b1 <- bracketsGrob(10/25, 0, 0.005, 0, h = 0.05)
b2 <- bracketsGrob(20/25, 0, 10/25, 0, h = 0.05)
b3 <- bracketsGrob(0.995, 0, 20/25, 0, h = 0.05)

bracket_labels <- tibble(x = c(5.5, 15.5, 23), #where to put labels on x-axis
                         y = rep(-1.75, 3), #how far below bracket to place labels
                         label = c("C1", "C2", "N")) #labels for groups of variables
herring.cor.p <-
  herring.cor.p +
  coord_cartesian(ylim = c(1,25), clip = "off") + #allows brackets to show up off plot
  theme(
    axis.text = element_blank(), #turns off axis labels
    axis.ticks = element_blank(), #turns off axis ticks
    plot.margin = unit(c(1,1,1.5,1), "lines") #padds margin to make room
  ) +
  annotation_custom(b1) +
  annotation_custom(b2) +
  annotation_custom(b3) +
  geom_text(data = bracket_labels, inherit.aes = FALSE, aes(x=x, y=y, label = label))
herring.cor.p
```
## PCA and PLS-DA for Red Herring

```{r}
herring.pcalr <- pca_lr(herring, -group, group, predI = 2, CV = 7)
herring.pca <- herring.pcalr$pca
herring.pls <- opls(select(herring, -group), null$group, fig.pdfC = "none", permI = 500, predI = 2)
```

## Create plots for Null

```{r message=FALSE, warning=FALSE}
herring.pca.p <-
  plot_pca(herring.pca, null$group, annotate = "subtitle") +
  labs(title = NULL,
       subtitle =TeX(glue("$R^2_Y = {signif(herring.pcalr$mod.stats$R2Y, 2)}$; $p = {signif(herring.pcalr$mod.stats$p.value, 2)}$"))) +
  scale_color_discrete_qualitative(palette = "Dark 3")

herring.pls.p <-
  plot_plsda(herring.pls, annotate = "subtitle") +
  scale_color_discrete_qualitative(palette = "Dark 3")
herring.pca.p
herring.pls.p
```

# Make figure

## Three Scenarios
Put Null, Control, and Needle together to make a figure.

```{r}
fix_theme_cor <-
  theme(plot.subtitle = element_text(hjust = 0.5))

fix_theme_score <-
  theme(legend.position = "none",
        plot.title = element_blank(),
        panel.grid = element_blank())
```

```{r}
library(patchwork)
# null.cor.p + control.cor.p + needle.cor.p + plot_layout(guides = "collect")
# guides = "collect" currently broken with R 4.0.0, workaround until patchwork update:
null.cor.p <- null.cor.p + theme(legend.position = "none")
control.cor.p <- control.cor.p + theme(legend.position = "none")

(null.cor.p + control.cor.p + needle.cor.p & fix_theme_cor) /
  (null.pca.p + control.pca.p + needle.pca.p & fix_theme_score) /
  (null.pls.p + control.pls.p + needle.pls.p & fix_theme_score) +
  plot_annotation(tag_levels = "a", tag_prefix = "(", tag_suffix = ")")
ggsave(here("out", "three-scenarios.png"), width = 10, height = 9)
```

# Biplots

Make biblots for each scenario for supplemental.

```{r}
y = rep(c("a", "b"), each = 10)
mytheme <- theme_bw() + theme(panel.grid = element_blank(),
                 text = element_text(size = 14))
```


## No predictors bi-plot


```{r}
null.pca.p <- my_biplot(null.pca, y)
null.pca.stats <- get_modelinfo(null.pca)
```


```{r}
r2_1 <-pca_lr(null, -group, group, reg_pcs = 1, predI = 2, CV = 7)$mod.stats$R2Y
r2_2 <-pca_lr(null, -group, group, reg_pcs = 2, predI = 2, CV = 7)$mod.stats$R2Y
xlab <- TeX(glue("Comp. 1 ($R^2_X = {null.pca.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {round(r2_1*100, 2)}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {null.pca.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {round(r2_2*100, 2)}%$)"))

null.pca.p <- 
  null.pca.p +
  labs(x = xlab,
       y = ylab) +
  scale_color_manual(values = c("N" = "grey30")) +
  mytheme 
null.pca.p
ggsave(here("out", "supplemental", "none pca biplot.png"), scale = 1.5)
```
```{r}
null.pls.p <- my_biplot(null.pls)
null.pls.stats <- get_modelinfo(null.pls)
```


```{r}
xlab <- TeX(glue("Comp. 1 ($R^2_X = {null.pls.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {null.pls.stats$axis_stats[1,3] * 100}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {null.pls.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {null.pls.stats$axis_stats[2,3] * 100}%$)"))

null.pls.p <- 
  null.pls.p +
  labs(x = xlab,
       y = ylab) +
  scale_color_manual(values = c("N" = "grey30")) +
  mytheme 
null.pls.p
ggsave(here("out", "supplemental", "none pls biplot.png"), scale = 1.5)
```

## Apparent predictors
```{r}
control.pca.p <- my_biplot(control.pca, group = y)
control.pca.stats <- get_modelinfo(control.pca)
```

```{r}
r2_1 <-pca_lr(control, -group, group, reg_pcs = 1, predI = 2, CV = 7)$mod.stats$R2Y
r2_2 <-pca_lr(control, -group, group, reg_pcs = 2, predI = 2, CV = 7)$mod.stats$R2Y
xlab <- TeX(glue("Comp. 1 ($R^2_X = {control.pca.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {round(r2_1*100, 2)}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {control.pca.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {round(r2_2*100, 2)}%$)"))

control.pca.p <- 
  control.pca.p +
  labs(x = xlab,
       y = ylab) +
  scale_color_manual(values = c("D1" = "#9B0000", "D2" = "#FF3939", "N" = "grey30")) +
  mytheme 
control.pca.p
ggsave(here("out", "supplemental", "apparent pca biplot.png"), scale = 1.5)
```
```{r}
control.pls.p <- my_biplot(control.pls)
control.pls.stats <- get_modelinfo(control.pls)
```

```{r}
xlab <- TeX(glue("Comp. 1 ($R^2_X = {control.pls.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {control.pls.stats$axis_stats[1,3] * 100}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {control.pls.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {control.pls.stats$axis_stats[2,3] * 100}%$)"))

control.pls.p <- 
  control.pls.p +
  labs(x = xlab,
       y = ylab) +
    scale_color_manual(values = c("D1" = "#9B0000", "D2" = "#FF3939", "N" = "grey30")) +
  mytheme 
control.pls.p
ggsave(here("out", "supplemental", "apparent pls biplot.png"), scale =1.5)
```
## Hidden predictors
```{r}
needle.pca.p <- my_biplot(needle.pca, group = y)
needle.pca.stats <- get_modelinfo(needle.pca)
```

```{r}
r2_1 <-pca_lr(needle, -group, group, reg_pcs = 1, predI = 2, CV = 7)$mod.stats$R2Y
r2_2 <-pca_lr(needle, -group, group, reg_pcs = 2, predI = 2, CV = 7)$mod.stats$R2Y
xlab <- TeX(glue("Comp. 1 ($R^2_X = {needle.pca.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {round(r2_1*100, 2)}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {needle.pca.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {round(r2_2*100, 2)}%$)"))

needle.pca.p <- 
  needle.pca.p +
  labs(x = xlab,
       y = ylab) +
    scale_color_manual(values = c("C1" = "#082568", "C2" = "#5777C0", "D" = "#9B0000")) +
  mytheme 
needle.pca.p
ggsave(here("out", "supplemental", "hidden pca biplot.png"), scale = 1.5)
```


```{r}
needle.pls.p <- my_biplot(needle.pls)
needle.pls.stats <- get_modelinfo(needle.pls)
```

```{r}
xlab <- TeX(glue("Comp. 1 ($R^2_X = {needle.pls.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {needle.pls.stats$axis_stats[1,3] * 100}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {needle.pls.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {needle.pls.stats$axis_stats[2,3] * 100}%$)"))


needle.pls.p <- 
  needle.pls.p +
  labs(x = xlab,
       y = ylab) +
      scale_color_manual(values = c("C1" = "#082568", "C2" = "#5777C0", "D" = "#9B0000")) +
  mytheme
needle.pls.p
ggsave(here("out", "supplemental", "hidden pls biplot.png"), scale = 1.5)
```

## Combine Biplots

I think I want 4 panels, PCA on the top, PLS on the bottom, apparent on the left, hidden on the right (to match figure 3)

```{r fig.height=7, fig.width=7.5}
(control.pca.p + labs(subtitle = "PCA", title = "Apparent Predictors"))+
  (needle.pca.p + labs(subtitle = "PCA", title = "Hidden Predictors")) +
  (control.pls.p + labs(subtitle = "PLS")) +
  (needle.pls.p + labs(subtitle = "PLS")) & scale_shape_manual(values = c("a" = "circle", "b" = "triangle open")) &
  theme(text = element_text(size=10),
        legend.position = "none", plot.title = element_text(hjust=0.5, vjust = 9)) &
  plot_annotation(tag_levels = "a", tag_prefix = "(", tag_suffix = ")")
ggsave(here("out", "simulation biplots.png"))
```
