==> R/6-simulation-figures.Rmd <==
---
title: "PCA vs. PLS figures"
author: "Eric R. Scott"
output: html_notebook
---

This generates figure 2 showing the correlation structure and results from PCA-LR and PLS-DA for example datasets from the three scenarios.

**IMPORTANT:** The random number generator in R was updated in version 3.6.0.  For this to be fully reproducible, run with version > 3.6.0

```{r setup, message=FALSE, warning=FALSE}
library(tidyverse)
library(ropls) #for opls(), which does pca and pls-da
library(here)
library(glue)
library(latex2exp)
library(patchwork)
library(colorspace)

source(here("R", "ropls_plots.R"))
source(here("R", "ropls_helpers.R"))
source(here("R", "pcr.R"))
```


# Read in datasets
I'll read in the datasets from "3-generating-datasets.Rmd" and pick individual datasets to make figures

```{r}
none.list <- read_rds(here::here("data", "none.rds"))
apparent.list <- read_rds(here::here("data", "apparent.rds"))
hidden.list <- read_rds(here::here("data", "hidden.rds"))
```

Sample one dataset as an example

```{r}
# set.seed(719)
set.seed(720)
none <- sample(none.list, 1)[[1]]
hidden <- sample(hidden.list, 1)[[1]]
apparent <- sample(apparent.list, 1)[[1]]
```

# "No Predictors"

## Correlation heatmap for "no predictors"

```{r}
none.cor <-
  none %>% 
  select(-group) %>% 
  cor() %>%
  as_tibble(rownames = "row") %>% 
  gather(-row, key = col, value = cor)

none.cor.p <-
  ggplot(none.cor, aes(x = col, y = row, fill = cor)) +
  geom_tile() +
  scale_fill_continuous_diverging(palette = "Blue-Red 3", limits = c(-1,1)) +
  labs(x = NULL, y = NULL, subtitle = '"No predictors"', fill = "r") +
  theme_minimal() +
  theme(legend.position = "right")

# Replace axis labels with curly braces to improve readability:
b <- bracketsGrob(0.995, 0, 0.005, 0, h = 0.05)

bracket_labels <- tibble(x = 13, y = -1.75, label = "N")

none.cor.p <-
  none.cor.p +
  coord_cartesian(ylim = c(1,25), clip = "off") + #allows brackets to show up off plot
  theme(
    axis.text = element_blank(), #turns off axis labels
    axis.ticks = element_blank(), #turns off axis ticks
    plot.margin = unit(c(1,1,1.5,1), "lines") #padds margin to make room
  ) +
  annotation_custom(b) +
  geom_text(data = bracket_labels, inherit.aes = FALSE, aes(x=x, y=y, label = label))

none.cor.p
```


## PCA and PLS-DA for "No Predictors"

```{r}
none.pcalr <- pca_lr(none, -group, group, predI = 2, CV = 7)
none.pca <- none.pcalr$pca
none.pls <- opls(select(none, -group), none$group, fig.pdfC = "none", permI = 500, predI = 2)
```

## Create plots for "No Predictors"

```{r message=FALSE, warning=FALSE}
none.pca.p <-
  plot_pca(none.pca, none$group, annotate = "caption") +
  scale_color_discrete_qualitative(palette = "Dark 3") +
  labs(caption =TeX(glue("$R^2_Y = {signif(none.pcalr$mod.stats$R2Y, 2)}$; $p = {signif(none.pcalr$mod.stats$p.value, 2)}$"))) +
  theme_classic() +
  theme(plot.title = element_blank(),
        plot.caption.position = "panel",
        plot.caption = element_text(vjust = 25, hjust = 0.02))


none.pls.p <-
  plot_plsda(none.pls, annotate = "caption") +
  scale_color_discrete_qualitative(palette = "Dark 3") +
  theme_classic() +
  theme(plot.title = element_blank(),
        plot.caption.position = "panel",
        plot.caption = element_text(vjust = 25, hjust = 0.02))
none.pca.p
none.pls.p
```


# "Hidden Predictors"

## Correlation heatmap for "hidden predictors"

```{r}
hidden.cor <-
  hidden %>% 
  select(-group) %>% 
  cor() %>%
  as_tibble(rownames = "row") %>%  
  gather(-row, key = col, value = cor)

hidden.cor.p <-
  ggplot(hidden.cor, aes(x = col, y = row, fill = cor)) +
  geom_tile() +
  scale_fill_continuous_diverging(palette = "Blue-Red 3", limits = c(-1,1)) +
  labs(x = NULL, y = NULL, subtitle = '"Hidden predictors"', fill = "r") +
  theme_minimal() +
  theme(legend.position = "right")

b1 <- bracketsGrob(10/25, 0, 0.005, 0, h = 0.05)
b2 <- bracketsGrob(20/25, 0, 10/25, 0, h = 0.05)
b3 <- bracketsGrob(0.995, 0, 20/25, 0, h = 0.05)

bracket_labels <- tibble(x = c(5.5, 15.5, 23), #where to put labels on x-axis
                         y = rep(-1.75, 3), #how far below bracket to place labels
                         label = c("C1", "C2", "D")) #labels for groups of variables

hidden.cor.p <-
  hidden.cor.p +
  coord_cartesian(ylim = c(1,25), clip = "off") + #allows brackets to show up off plot
  theme(
    axis.text = element_blank(), #turns off axis labels
    axis.ticks = element_blank(), #turns off axis ticks
    plot.margin = unit(c(1,1,1.5,1), "lines") #pads margin to make room
  ) +
  annotation_custom(b1) +
  annotation_custom(b2) +
  annotation_custom(b3) +
  geom_text(data = bracket_labels, inherit.aes = FALSE, aes(x=x, y=y, label = label))
hidden.cor.p
```

## PCA and PLS-DA for "hidden predictors"

The PCA score plot still shows no real separation because the discriminating variables do not contribute much to the total covariation in the data.  The PLS-DA plot shows very strong separation, despite the differences between the groups being only due to 5 out of 25 variables.  The PLS-DA model performs well, with high R^2^ and Q^2^ values

```{r include=FALSE}
hidden.pcalr <- pca_lr(hidden, -group, group, predI = 2, CV = 7)
hidden.pca <- hidden.pcalr$pca
hidden.pls <- opls(select(hidden, -group), hidden$group, fig.pdfC = "none", predI = 2, permI = 500)
```

## Plots for "hidden predictors"

```{r message=FALSE, warning=FALSE}
hidden.pca.p <-
  plot_pca(hidden.pca, hidden$group, annotate = "caption") +
  scale_color_discrete_qualitative(palette = "Dark 3") +
  labs(caption =TeX(glue("$R^2_Y = {signif(hidden.pcalr$mod.stats$R2Y, 2)}$; $p = {signif(hidden.pcalr$mod.stats$p.value, 2)}$"))) +
  theme_classic() +
  theme(plot.title = element_blank(),
        plot.caption.position = "panel",
        plot.caption = element_text(vjust = 25, hjust = 0.02))

hidden.pls.p <-
  plot_plsda(hidden.pls, annotate = "caption") +
  scale_color_discrete_qualitative(palette = "Dark 3") +
  theme_classic() +
  theme(plot.title = element_blank(),
        plot.caption.position = "panel",
        plot.caption = element_text(vjust = 25, hjust = 0.02))
hidden.pca.p
hidden.pls.p
```


# "Apparent Predictors"

## Correlation heatmap for "Apparent Predictors"

```{r}
apparent.cor <-
  apparent %>% 
  select(-group) %>% 
  cor() %>%
  as_tibble(rownames = "row") %>% 
  gather(-row, key = col, value = cor)

apparent.cor.p <-
  ggplot(apparent.cor, aes(x = col, y = row, fill = cor)) +
  geom_tile() +
  scale_fill_continuous_diverging(palette = "Blue-Red 3", limits = c(-1,1)) +
  labs(x = NULL, y = NULL, subtitle = '"Apparent predictors"', fill = "r") +
  theme_minimal()

b1 <- bracketsGrob(10/25, 0, 0.005, 0, h = 0.05)
b2 <- bracketsGrob(20/25, 0, 10/25, 0, h = 0.05)
b3 <- bracketsGrob(0.995, 0, 20/25, 0, h = 0.05)

bracket_labels <- tibble(x = c(5.5, 15.5, 23), #where to put labels on x-axis
                         y = rep(-1.75, 3), #how far below bracket to place labels
                         label = c("D1", "D2", "N")) #labels for groups of variables
apparent.cor.p <-
  apparent.cor.p +
  coord_cartesian(ylim = c(1,25), clip = "off") + #allows brackets to show up off plot
  theme(
    axis.text = element_blank(), #turns off axis labels
    axis.ticks = element_blank(), #turns off axis ticks
    plot.margin = unit(c(1,1,1.5,1), "lines") #padds margin to make room
  ) +
  annotation_custom(b1) +
  annotation_custom(b2) +
  annotation_custom(b3) +
  geom_text(data = bracket_labels, inherit.aes = FALSE, aes(x=x, y=y, label = label))
apparent.cor.p
```

## PCA and PLS-DA for "Apparent Predictors"

```{r include=FALSE}
apparent.pcalr <- pca_lr(apparent, -group, group,  CV = 7, predI = 2) 

apparent.pca <- apparent.pcalr$pca
apparent.pls <- opls(select(apparent, -group), apparent$group, fig.pdfC = "none", predI = 2, permI = 500)
```

## Plots for "Apparent Predictors"

```{r message=FALSE, warning=FALSE}
apparent.pca.p <-
  plot_pca(apparent.pca, apparent$group, annotate = "caption") +
  scale_color_discrete_qualitative(palette = "Dark 3") +
  labs(caption = TeX(glue("$R^2_Y = 0.99$; $p < 0.001$"))) +
  theme_classic() +
  theme(plot.title = element_blank(),
        plot.caption.position = "panel",
        plot.caption = element_text(vjust = 25, hjust = 0.02))
  
apparent.pls.p <-
  plot_plsda(apparent.pls, annotate = "caption") +
  scale_color_discrete_qualitative(palette = "Dark 3") +
  theme_classic() +
  theme(plot.title = element_blank(),
        plot.caption.position = "panel",
        plot.caption = element_text(vjust = 25, hjust = 0.02))
apparent.pca.p
apparent.pls.p
```

# Make figure
## Three Scenarios
Put three scenarios together to make a figure.


```{r fig.width=9.5, fig.height=2.833}
r1 <-
  (none.cor.p +       annotate("text", x = -Inf, y = Inf, label = "(a)", hjust = -0.2, vjust = 1.2) |
     apparent.cor.p + annotate("text", x = -Inf, y = Inf, label = "(b)", hjust = -0.2, vjust = 1.2)|
     hidden.cor.p +   annotate("text", x = -Inf, y = Inf, label = "(c)", hjust = -0.2, vjust = 1.2)) +
  plot_layout(guides = "collect") &
  #center subtitles
  theme(plot.subtitle = element_text(hjust = 0.5))
r1
```


```{r fig.width=9.5, fig.height=2.833}
r2 <-
  (none.pca.p +       annotate("text", x = -Inf, y = Inf, label = "(d)", hjust = -0.2, vjust = 1.2)| 
     apparent.pca.p + annotate("text", x = -Inf, y = Inf, label = "(e)", hjust = -0.2, vjust = 1.2) | 
     hidden.pca.p +   annotate("text", x = -Inf, y = Inf, label = "(f)", hjust = -0.2, vjust = 1.2)) &
  scale_y_continuous(expand = expansion(mult = c(0.12, 0.05))) &
  theme(legend.position = "none")
r2
```


```{r fig.width=9.5, fig.height=2.833}
r3 <-
  (none.pls.p +       annotate("text", x = -Inf, y = Inf, label = "(g)", hjust = -0.2, vjust = 1.2)| 
     apparent.pls.p + annotate("text", x = -Inf, y = Inf, label = "(h)", hjust = -0.2, vjust = 1.2)|
     hidden.pls.p +   annotate("text", x = -Inf, y = Inf, label = "(i)", hjust = -0.2, vjust = 1.2)) &
  scale_y_continuous(expand = expansion(mult = c(0.12, 0.05))) &
  theme(legend.position = "none")
r3
```
Can't get tags inside of plots with patchwork, using `annotate()` instead

```{r fig.width=9.5, fig.height=8.5}
(r1/r2/r3) 
```


Gonna make a version with no tags and just add them in the pdf I guess?

```{r fig.width=9.5, fig.height=8.5}
r1/r2/r3 + plot_annotation(title = "Fig. 4")

ggsave(here("out", "three-scenarios.png"), width = 9.5, height = 8.5)
ggsave(here("out", "three-scenarios.eps"), width = 9.5, height = 8.5)
ggsave(here("out", "three-scenarios.pdf"), width = 9.5, height = 8.5)
```

# Biplots

Make biplots for each scenario for supplemental.

```{r}
y = rep(c("a", "b"), each = 10)
mytheme <- theme_bw() + theme(panel.grid = element_blank(),
                 text = element_text(size = 14))
```


## No predictors bi-plot


```{r}
none.pca.bp <- my_biplot(none.pca, y)
none.pca.stats <- get_modelinfo(none.pca)
```


```{r}
r2_1 <-pca_lr(none, -group, group, reg_pcs = 1, predI = 2, CV = 7)$mod.stats$R2Y
r2_2 <-pca_lr(none, -group, group, reg_pcs = 2, predI = 2, CV = 7)$mod.stats$R2Y
xlab <- TeX(glue("Comp. 1 ($R^2_X = {none.pca.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {round(r2_1*100, 2)}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {none.pca.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {round(r2_2*100, 2)}%$)"))

none.pca.bp <- 
  none.pca.bp +
  labs(x = xlab,
       y = ylab) +
  scale_color_manual(values = c("N" = "grey30")) +
  mytheme 
none.pca.bp
```
```{r}
none.pls.bp <- my_biplot(none.pls)
none.pls.stats <- get_modelinfo(none.pls)
```


```{r}
xlab <- TeX(glue("Comp. 1 ($R^2_X = {none.pls.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {none.pls.stats$axis_stats[1,3] * 100}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {none.pls.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {none.pls.stats$axis_stats[2,3] * 100}%$)"))

none.pls.bp <- 
  none.pls.bp +
  labs(x = xlab,
       y = ylab) +
  scale_color_manual(values = c("N" = "grey30")) +
  mytheme 
none.pls.bp
```

## Apparent predictors
```{r}
apparent.pca.bp <- my_biplot(apparent.pca, group = y)
apparent.pca.stats <- get_modelinfo(apparent.pca)
```

```{r}
r2_1 <-pca_lr(apparent, -group, group, reg_pcs = 1, predI = 2, CV = 7)$mod.stats$R2Y
r2_2 <-pca_lr(apparent, -group, group, reg_pcs = 2, predI = 2, CV = 7)$mod.stats$R2Y
xlab <- TeX(glue("Comp. 1 ($R^2_X = {apparent.pca.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {round(r2_1*100, 2)}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {apparent.pca.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {round(r2_2*100, 2)}%$)"))

apparent.pca.bp <- 
  apparent.pca.bp +
  labs(x = xlab,
       y = ylab) +
  scale_color_manual(values = c("D1" = "#9B0000", "D2" = "#FF3939", "N" = "grey30")) +
  mytheme 
apparent.pca.bp
```
```{r}
apparent.pls.bp <- my_biplot(apparent.pls)
apparent.pls.stats <- get_modelinfo(apparent.pls)
```

```{r}
xlab <- TeX(glue("Comp. 1 ($R^2_X = {apparent.pls.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {apparent.pls.stats$axis_stats[1,3] * 100}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {apparent.pls.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {apparent.pls.stats$axis_stats[2,3] * 100}%$)"))

apparent.pls.bp <- 
  apparent.pls.bp +
  labs(x = xlab,
       y = ylab) +
    scale_color_manual(values = c("D1" = "#9B0000", "D2" = "#FF3939", "N" = "grey30")) +
  mytheme 
apparent.pls.bp
```
## Hidden predictors
```{r}
hidden.pca.bp <- my_biplot(hidden.pca, group = y)
hidden.pca.stats <- get_modelinfo(hidden.pca)
```

```{r}
r2_1 <-pca_lr(hidden, -group, group, reg_pcs = 1, predI = 2, CV = 7)$mod.stats$R2Y
r2_2 <-pca_lr(hidden, -group, group, reg_pcs = 2, predI = 2, CV = 7)$mod.stats$R2Y
xlab <- TeX(glue("Comp. 1 ($R^2_X = {hidden.pca.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {round(r2_1*100, 2)}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {hidden.pca.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {round(r2_2*100, 2)}%$)"))

hidden.pca.bp <- 
  hidden.pca.bp +
  labs(x = xlab,
       y = ylab) +
    scale_color_manual(values = c("C1" = "#082568", "C2" = "#5777C0", "D" = "#9B0000")) +
  mytheme 
hidden.pca.bp
```


```{r}
hidden.pls.bp <- my_biplot(hidden.pls)
hidden.pls.stats <- get_modelinfo(hidden.pls)
```

```{r}
xlab <- TeX(glue("Comp. 1 ($R^2_X = {hidden.pls.stats$axis_stats[1,1] * 100}%$, $R^2_Y = {hidden.pls.stats$axis_stats[1,3] * 100}%$)"))
ylab <- TeX(glue("Comp. 2 ($R^2_X = {hidden.pls.stats$axis_stats[2,1] * 100}%$, $R^2_Y = {hidden.pls.stats$axis_stats[2,3] * 100}%$)"))


hidden.pls.bp <- 
  hidden.pls.bp +
  labs(x = xlab,
       y = ylab) +
      scale_color_manual(values = c("C1" = "#082568", "C2" = "#5777C0", "D" = "#9B0000")) +
  mytheme
hidden.pls.bp
```


## Combine Biplots

I think I want 4 panels, PCA on the top, PLS on the bottom, apparent on the left, hidden on the right (to match figure 3)

```{r fig.height=6, fig.width=6.5}
biplots <-
  (apparent.pca.bp + labs(subtitle = "a) PCA, Apparent Predictors"))+
  (hidden.pca.bp + labs(subtitle = "b) PCA, Hidden Predictors")) +
  (apparent.pls.bp + labs(subtitle = "c) PLS, Apparent Predictors")) +
  (hidden.pls.bp + labs(subtitle = "d) PLS, Hidden Predictors")) &
  scale_shape_manual(values = c("a" = "circle", "b" = "triangle open")) &
  scale_y_continuous(expand = expansion(mult = c(0.05, 0.15))) &
  theme_classic() &
  theme(text = element_text(size=10),
        legend.position = "none",
        plot.subtitle = element_text(hjust=0.1, vjust = -8)) 
biplots + plot_annotation(title = "Fig. 5")
ggsave(here("out", "simulation biplots.png"), height = 84*2, width = 84*2, units = "mm")
ggsave(here("out", "simulation biplots.pdf"), height = 84*2, width = 84*2, units = "mm")
```
